+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+# 1.1.0 (2025-01-04)
+
+### Added
+
+- `EnYupSchema` a schema with also `pick` and `omit`
+
+### Changed
+
+- `getSchemaByType` and `getNamedSchema` return `EnYupSchema`
+
 # 1.0.0 (2024-12-30)
 
 ### Features